Extract some shared Okta functionality
This is to prepare us for new views for logging in users to Okta for organizations.
